backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
celery[redis]==5.4.0
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
websockets==14.1
aiohttp==3.11.11
python-telegram-bot==21.9
stripe==11.4.0
numpy==2.2.1
pandas==2.2.3
ta==0.11.0
python-dotenv==1.0.1
loguru==0.7.3
pytest==8.3.4
pytest-asyncio==0.25.0
